# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/shops": {
      "get": {
        "operationId": "AdminShopsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "example": "PENDING",
              "type": "string",
              "enum": [
                "PENDING",
                "ACTIVE",
                "SUSPENDED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Ali Market",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 20,
              "example": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "{ items, total, page, limit, totalPages }"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Do‘konlar ro‘yxati (status filtr + sahifalash)",
        "tags": [
          "admin-shops"
        ]
      }
    },
    "/api/v1/admin/shops/{id}": {
      "get": {
        "operationId": "AdminShopsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminShopDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Do‘kon profili va asosiy statistikasi",
        "tags": [
          "admin-shops"
        ]
      }
    },
    "/api/v1/admin/shops/{id}/suspend": {
      "post": {
        "operationId": "AdminShopsController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "ACTIVE do‘konni vaqtincha to‘xtatish",
        "tags": [
          "admin-shops"
        ]
      }
    },
    "/api/v1/admin/shops/{id}/activate": {
      "post": {
        "operationId": "AdminShopsController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "SUSPENDED do‘konni qayta faollashtirish",
        "tags": [
          "admin-shops"
        ]
      }
    },
    "/api/v1/admin/shops/{id}/feature": {
      "post": {
        "operationId": "AdminShopsController_feature",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureShopDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Do‘konni bosh sahifada ko‘rsatish/yashirish",
        "tags": [
          "admin-shops"
        ]
      }
    },
    "/api/v1/admin/shops/{id}/tariffs": {
      "patch": {
        "operationId": "AdminShopsController_updateTariffs",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShopTariffsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Do‘konning Elchi yetkazish tariflarini sozlash/sinxronlash",
        "tags": [
          "admin-shops"
        ]
      }
    },
    "/api/v1/admin/shops/{id}/approve": {
      "post": {
        "operationId": "AdminShopsController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Do‘konni tasdiqlash (active + user active + default ombor)",
        "tags": [
          "admin-shops"
        ]
      }
    },
    "/api/v1/admin/shops/{id}/reject": {
      "post": {
        "operationId": "AdminShopsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectShopDto"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Do‘konni rad etish (+ sotuvchiga xabar)",
        "tags": [
          "admin-shops"
        ]
      }
    },
...
  },
  "info": {
    "title": "Elchi Marketplace API",
    "description": "Marketplace API — MVP (batafsil: docs/API_CONTRACT.md)",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": ""
          },
          "error": {
            "type": "string",
            "example": "phone +998XXXXXXXXX formatida bo‘lishi kerak"
          }
        },
        "required": [
          "field",
          "error"
        ]
      },
      "AuthErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Telefon yoki parol xato"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "password must be longer than or equal to 8 characters"
                ]
              }
            ]
          },
          "errorCode": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetailDto"
            }
          }
        },
        "required": [
          "statusCode",
          "message",
          "errorCode"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "AdminShopDetailStatsDto": {
        "type": "object",
        "properties": {
          "products": {
            "type": "number",
            "example": 12
          },
          "orders": {
            "type": "number",
            "example": 28
          },
          "warehouses": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "products",
          "orders",
          "warehouses"
        ]
      },
      "AdminShopDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "15"
          },
          "ownerUserId": {
            "type": "string",
            "example": "42"
          },
          "name": {
            "type": "string",
            "example": "Ali Market"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACTIVE",
              "SUSPENDED",
              "REJECTED"
            ]
          },
          "tariffHome": {
            "type": "number",
            "example": 25000
          },
          "tariffCenter": {
            "type": "number",
            "example": 15000
          },
          "stats": {
            "$ref": "#/components/schemas/AdminShopDetailStatsDto"
          }
        },
        "required": [
          "id",
          "ownerUserId",
          "name",
          "status",
          "tariffHome",
          "tariffCenter",
          "stats"
        ]
      },
      "FeatureShopDto": {
        "type": "object",
        "properties": {
          "featured": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "featured"
        ]
      },
      "UpdateShopTariffsDto": {
        "type": "object",
        "properties": {
          "tariffHome": {
            "type": "number",
            "example": 25000,
            "minimum": 1
          },
          "tariffCenter": {
            "type": "number",
            "example": 15000,
            "minimum": 1
          }
        },
        "required": [
          "tariffHome",
          "tariffCenter"
        ]
      },
      "RejectShopDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Hujjatlar to‘liq emas"
          }
        }
      },
...
    }
  }
}